+++ example/.env.example
@@ -11,8 +11,10 @@ DISPLAY_RUNTIME_URL=http://localhost:8003
 SERVER_RUNTIME_URL=http://localhost:8004
 
 # AI service configuration
-AI_UI_ENABLED=true
-AI_MODEL_ID=gpt-4o
+# If AI_UI_ENABLED is set to true, the AI service will be enabled and model id
+# and secret key must be provided.
+AI_UI_ENABLED=
+AI_MODEL_ID=
 AI_SECRET_KEY=
 
 # Content Element env variables; TCE_ prefix is required
