+++ packages/runtime/tce-server/config/ai.ts
@@ -1,3 +1,3 @@
 export const secretKey = process.env.AI_SECRET_KEY;
 export const modelId = process.env.AI_MODEL_ID;
-export const isEnabled = secretKey && modelId;
+export const isConfigured = process.env.AI_SECRET_KEY;
